requirements.txt
arrow==0.12.1
asn1crypto==0.24.0
backports.functools-lru-cache==1.5
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
Django==1.11.12
django-cors-headers==2.2.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
enum34==1.1.6
flake8==2.4.1
gunicorn==19.7.0
idna==2.5
ipaddress==1.0.22
mccabe==0.3.1
mysqlclient==1.3.12
pep8==1.7.1
pkg-resources==0.0.0
pycparser==2.19
pyflakes==0.8.1
PyJWT==1.6.4
python-dateutil==2.7.2
python-dotenv==0.9.1
pytz==2018.4
requests==2.18.0
six==1.11.0
stripe==1.84.0
urllib3==1.21.1
